Update dbn_sample to change behavior based on dynamic variables
Updates dbn_sample to use previous implementation if a dynamic variable depends on another dynamic variable. In release 1.4.0 dbn_sample was updated to calculate the index, j, upfront because asub2ind can introduce unwanted overhead and also preallocated events as a NaN array. In this previous release, the for ii = order_transition loop was added to identify the relationship between dynamic variables and its parents. Notably in the for ii = order_transition loop, the variable x was not updated. Now this is where the bug was introduced. If a dynamic variable was dependent on another dynamic variable (see unconventional glider model), xj = x(parents) would be equal for the element with the dynamic variable dependence. This would results in asub2ind(rj, xj) returning a negative value, which would create an error when indexing N_transition{ii}(:, j(ii)). Since the uncorrelated conventional models transition networks do not have any dynamic variables not dependent on another dynamic variable, this bug was not identified in release 1.4.0. For this release, the bug was addressed by determining if any of the dynamic variables depend on another dynamic variable. This determines if we can calculate the index, j, upfront or via each iterate of t. If there is a dependence, it will sample the model similar to Release 1.3.0 where the events matrix was also preallocated as an empty array